Fix job check-in re-register

+++ propius/controller/client_manager/__main__.py
@@ -23,8 +23,8 @@ async def serve(gconfig, cm_id: int, logger: Propius_logger):
         logger.print(f"=====Client manager shutting down=====", Msg_level.WARNING)
         client_manager.cm_monitor.report(client_manager.cm_id)
         client_manager.client_db_portal.flushdb()
-        client_manager.temp_client_db_portal.flushdb()
-
+        if client_manager.sched_mode == "online":
+            client_manager.temp_client_db_portal.flushdb()
         try:
             client_assign_task.cancel()
             plot_task.cancel()
